68 brankbao output optiimize

+++ plugins/file.go
@@ -51,6 +51,7 @@ type FileCopyOutputs struct {
 
 type FileCopyOutput struct {
 	CallBackParameter
+	Result
 	Guid   string `json:"guid,omitempty"`
 	Detail string `json:"detail,omitempty"`
 }
@@ -67,31 +68,23 @@ func (action *FileCopyAction) ReadParam(param interface{}) (interface{}, error) 
 	return inputs, nil
 }
 
-func (action *FileCopyAction) CheckParam(input interface{}) error {
-	inputs, ok := input.(FileCopyInputs)
-	if !ok {
-		return fmt.Errorf("fileCopyAtion:input type=%T not right", input)
+func (action *FileCopyAction) CheckParam(input FileCopyInput) error {
+	if input.EndPoint == "" {
+		return errors.New("EndPoint is empty")
+	}
+	if input.Target == "" {
+		return errors.New("Target is empty")
+	}
+	if input.DestinationPath == "" {
+		return errors.New("DestinationPath is empty")
 	}
 
-	for _, input := range inputs.Inputs {
-		if input.EndPoint == "" {
-			return errors.New("EndPoint is empty")
-		}
-		if input.Target == "" {
-			return errors.New("Target is empty")
-		}
-		if input.DestinationPath == "" {
-			return errors.New("DestinationPath is empty")
-		}
-
-		// if input.SecretKey == "" {
-		// 	return errors.New("SecretKey is empty")
-		// }
-		// if input.AccessKey == "" {
-		// 	return errors.New("AccessKey is empty")
-		// }
-
-	}
+	// if input.SecretKey == "" {
+	// 	return errors.New("SecretKey is empty")
+	// }
+	// if input.AccessKey == "" {
+	// 	return errors.New("AccessKey is empty")
+	// }
 
 	return nil
 }
@@ -125,19 +118,37 @@ func changeDirecoryOwner(input *FileCopyInput) error {
 	return nil
 }
 
-func (action *FileCopyAction) copyFile(input *FileCopyInput) (*FileCopyOutput, error) {
-	output := FileCopyOutput{}
-	fileName, err := downloadS3File(input.EndPoint, "access_key", "secret_key")
+func (action *FileCopyAction) copyFile(input *FileCopyInput) (output FileCopyOutput, err error) {
+	defer func() {
+		output.Guid = input.Guid
+		output.CallBackParameter.Parameter = input.CallBackParameter.Parameter
+		if err == nil {
+			output.Result.Code = RESULT_CODE_SUCCESS
+		} else {
+			output.Result.Code = RESULT_CODE_ERROR
+			output.Result.Message = err.Error()
+		}
+	}()
+
+	err = action.CheckParam(*input)
 	if err != nil {
-		logrus.Errorf("CopyFile downloads3 file error=%v", err)
-		return &output, fmt.Errorf("CopyFile downloads3 file error=%v", err)
+		return output, err
 	}
+
+	fileName, er := downloadS3File(input.EndPoint, "access_key", "secret_key")
+	if er != nil {
+		logrus.Errorf("CopyFile downloads3 file error=%v", er)
+		err = fmt.Errorf("CopyFile downloads3 file error=%v", er)
+		return output, err
+	}
+
 	input.DestinationPath = buildFileDestinationPath(input.EndPoint, input.DestinationPath)
 
-	savePath, err := saveFileToSaltMasterBaseDir(fileName)
+	savePath, er := saveFileToSaltMasterBaseDir(fileName)
 	os.Remove(fileName)
-	if err != nil {
-		return &output, fmt.Errorf("saveFileToSaltMasterBaseDir meet error=%v", err)
+	if er != nil {
+		err = fmt.Errorf("saveFileToSaltMasterBaseDir meet error=%v", err)
+		return output, err
 	}
 
 	//copy file
@@ -145,34 +156,34 @@ func (action *FileCopyAction) copyFile(input *FileCopyInput) (*FileCopyOutput, e
 	_, err = CallSaltApi("https://127.0.0.1:8080", *copyRequest)
 	os.Remove(savePath)
 	if err != nil {
-		return nil, err
+		return output, err
 	}
 
 	md5SumRequest, _ := action.deriveMd5SumRequest(input)
 	md5sum, err := CallSaltApi("https://127.0.0.1:8080", *md5SumRequest)
 	if err != nil {
-		return nil, err
+		return output, err
 	}
 
+	var unpackRequest *SaltApiRequest
 	if input.Unpack == "true" {
-		unpackRequest, err := action.deriveUnpackRequest(input)
+		unpackRequest, err = action.deriveUnpackRequest(input)
 		if err != nil {
-			return nil, err
+			return output, err
 		}
 
 		if _, err = CallSaltApi("https://127.0.0.1:8080", *unpackRequest); err != nil {
-			return nil, err
+			return output, err
 		}
 		if input.FileOwner != "" {
-			if err := changeDirecoryOwner(input); err != nil {
-				return nil, err
+			if err = changeDirecoryOwner(input); err != nil {
+				return output, err
 			}
 		}
 	}
 
-	output.Guid = input.Guid
 	output.Detail = md5sum
-	return &output, nil
+	return output, err
 }
 
 func (action *FileCopyAction) deriveMd5SumRequest(input *FileCopyInput) (*SaltApiRequest, error) {
@@ -243,16 +254,16 @@ func (action *FileCopyAction) deriveUnpackRequest(input *FileCopyInput) (*SaltAp
 func (action *FileCopyAction) Do(input interface{}) (interface{}, error) {
 	files, _ := input.(FileCopyInputs)
 	outputs := FileCopyOutputs{}
+	var finalErr error
 	for _, file := range files.Inputs {
 		fileCopyOutput, err := action.copyFile(&file)
-		fileCopyOutput.CallBackParameter.Parameter = file.CallBackParameter.Parameter
-
 		if err != nil {
-			return nil, err
+			finalErr = err
 		}
-		outputs.Outputs = append(outputs.Outputs, *fileCopyOutput)
+
+		outputs.Outputs = append(outputs.Outputs, fileCopyOutput)
 	}
 
 	logrus.Infof("all files = %v are copied", files)
-	return &outputs, nil
+	return &outputs, finalErr
 }
